+++ jgclark.Dashboard/src/react/components/TasksFiltered.jsx
@@ -2,7 +2,7 @@
 //--------------------------------------------------------------------------
 // Dashboard React component to show an Indicator that a Filter has been applied and so some item(s) have been hidden.
 // Called by ItemRow component
-// Last updated 2025-12-01 for v2.3.2, @jgclark
+// Last updated 2025-12-04 for v2.3.3, @jgclark
 //--------------------------------------------------------------------------
 
 import React, { type Node } from 'react'
@@ -40,16 +40,24 @@ const TasksFiltered = ({ item, onToggleShowAll }: Props): Node => {
   }
 
   return (
+
     <div className="sectionItemRow messageItemRow clickTarget" id={item.ID}>
       {/* This empty span needed to mimic the StatusIcon line */}
       <span>
-        <div className="itemIcon">
-          <i id={item.ID} className="fa-regular fa-plus"></i>
-        </div>
+        {/* Check if this is the "Showing all N items..." message - if so, render empty StatusIcon */}
+        {item.itemType === 'offerToFilter' &&
+          <div className="emptyIcon" />
+        }
+        {/* Check if this is the "click to filter by priority" message - if so, render '+' StatusIcon */}
+        {item.itemType === 'filterIndicator' && 
+          <div className="itemIcon">
+            <i id={item.ID} className={'fa-regular fa-plus'} />
+          </div>
+        }
       </span>
       <div className="sectionItemContent sectionItem" onClick={(e) => handleLineClick(e)}>
         <span className="messageItemRow">
-          <i>{item?.message || '<no content>'}</i>
+          <i>{item.message || '<no content>'}</i>
         </span>
       </div>
     </div>